requirements.txt
streamlit
langchain
langchain-community
langchain-ollama
langgraph
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-fastembed
chromadb
langchain-chroma
fastembed
flashrank
pdfplumber
unstructured
rank_bm25
ragas
pydantic
loguru
requests
python-dotenv
tiktoken
pandas
